refactor: Set LZ4 HC compression level to minimum for faster caching
To prioritize a faster initial cover caching process, the LZ4 compression level has been set to the minimum (`LZ4HC_CLEVEL_MIN`).

Performance testing showed that this provides a significant speed improvement during cache creation. This change results in a ~4.5% increase in the average size per cached cover, but the faster startup time is considered a worthwhile trade-off for a better user experience.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>