feat: Check network connection before cover download
Adds a pre-flight check for an active Wi-Fi connection. This provides the user with an immediate "No network connection" message instead of waiting for a long timeout.

The check is disabled in the simulator to facilitate testing. Also fixes a related crash by increasing the size of the status message buffer.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>