feat(ui): Full-height hold animation in settings
The hold-to-activate animation for menu items was previously not as tall as the selector box itself.

This commit adjusts the rendering logic to make the animation bar fill the entire height of the selector. This provides a more polished and intuitive visual feedback to the user when holding a button on an actionable item.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>