feat(settings): Play sound on save/load state action
Adds a confirmation sound effect when the user presses 'A' to trigger a "Save state" or "Load state" action from the in-game settings menu.

Previously, there was no auditory feedback for this action, which was inconsistent with other UI interactions. This change provides immediate feedback by playing a sound when the action is initiated.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>