UI: Clarify description for Stabilization setting
Reworded the description for the "Stabilization" option in the settings menu to improve clarity and fit within the UI's constrained layout.

The new text is more concise and removes technical jargon. It also adds a recommendation to use the "Staggered" dither setting for the best visual results with this feature.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>